fix: Project structure and MessageList page
Also removed unused code and reformatted files
