+++ src/views/contact.hbs
@@ -1,0 +1,46 @@
+<html>
+
+<head>
+  <title>reCAPTCHA demo: Explicit render after an onload callback</title>
+  <script type="text/javascript">
+    var onloadCallback = function () {
+      grecaptcha.render('html_element', {
+        'sitekey': '6LcRi-AUAAAAADkzzqkKUigR69utt7ik07R93xUg'
+      });
+    };
+  </script>
+</head>
+
+<body>
+  <div class="fondecran">
+    <div class="form">
+      <h1 class="text">Message</h1>
+      <p class="text">Envoyez-nous un message pour nous faire part de toutes vos remarques, recommandations,
+        suggestions...</p>
+    </div>
+
+    <form action="/listeContacts" method="POST" enctype="multiparty/form-data" class="form">
+
+      <label class="labelText" for="nom">Nom</label>
+      <input class="input" type="text" name="nom" id="nom">
+
+      <label class="labelText" for="prenom">Prenom</label>
+      <input class="input" type="text" name="prenom" id="prenom">
+
+      <label class="labelText" for="mail">Email</label>
+      <input class="input" type="text" name="mail" id="mail">
+
+      <label class="labelText" for="message">Message</label>
+      <textarea class="input" name="message" id="message" cols="30" rows="10"></textarea>
+
+      <div class="input" id="html_element" required></div>
+      <br>
+      <script src="https://www.google.com/recaptcha/api.js?onload=onloadCallback&render=explicit" async defer required>
+      </script>
+      <button class="btn-light" id="send" type="submit">Envoyer</button>
+
+    </form>
+
+
+  </div>
+</body>
