+++ src/views/index.hbs
@@ -1,0 +1,70 @@
+<div class="fondecran">
+  <div id="description" class="view intro-2">
+    <div class="full-bg-img">
+      <div class="mask rgba-black-strong flex-center">
+        <div class="container">
+          <div class="white-text text-center wow fadeInUp" id="content">
+            <h2 class="text">La méditation pleine conscience</h2>
+            <h5 class="text">La pleine conscience est une expression désignant une attitude d'attention, de présence et
+              de conscience vigilante, qui peut être interne (sensations, pensées, émotions, actions, motivations, etc.)
+              ou externe (au monde environnant, bruits, objets, événements, etc.).</h5>
+          </div>
+        </div>
+      </div>
+    </div>
+  </div>
+
+  <div class="prettySlide">
+    <div id="myCarousel" class="carousel slide" data-ride="carousel">
+      <!-- Indicators -->
+      <ol class="carousel-indicators">
+        <li data-target="#myCarousel" data-slide-to="0" class="active"></li>
+        <li data-target="#myCarousel" data-slide-to="1"></li>
+        <li data-target="#myCarousel" data-slide-to="2"></li>
+        <li data-target="#myCarousel" data-slide-to="3"></li>
+      </ol>
+      <!-- Wrapper for slides -->
+      <div class="carousel-inner" role="listbox">
+        <div class="item active">
+          <img src="assets/img/pose.jpg" alt="Cheveux">
+        </div>
+        <div class="item">
+          <img src="assets/img/arbre2.jpg" alt="thé">
+        </div>
+        <div class="item">
+          <img src="assets/img/mer.jpg" alt="bulle">
+        </div>
+        <div class="item">
+          <img src="assets/img/cerveau.jpg" alt="arbre">
+        </div>
+      </div>
+      <!-- Left and right controls -->
+      <a class="left carousel-control" href="#myCarousel" role="button" data-slide="prev">
+        <span class="glyphicon glyphicon-chevron-left" aria-hidden="true"></span>
+        <span class="sr-only">Previous</span>
+      </a>
+      <a class="right carousel-control" href="#myCarousel" role="button" data-slide="next">
+        <span class="glyphicon glyphicon-chevron-right" aria-hidden="true"></span>
+        <span class="sr-only">Next</span>
+      </a>
+    </div>
+  </div>
+
+  <div id="description">
+    <div class="view intro-2">
+      <div class="full-bg-img">
+        <div class="mask rgba-black-strong flex-center">
+          <div class="container">
+            <p class="text" id="meditext">La méditation permet de développer cet état de pleine conscience. Elle
+              entraîne l'esprit à se libérer du flot de pensées pour se concentrer sur le moment présent plus longtemps
+              et plus profondément. Les bienfaits sont tels que la méditation en pleine conscience est de plus en plus
+              utilisée de façon clinique pour soigner des maladies mentales, et notament celles liées au stress. Mais de
+              manière générale, tous ceux qui la pratiquent ressentent un calme mental et un meilleur contrôle de leurs
+              émotions, pendant la méditation, mais aussi dans leur vie quotidienne.</p>
+          </div>
+        </div>
+      </div>
+    </div>
+  </div>
+
+</div>
