+++ src/views/layout.hbs
@@ -1,0 +1,37 @@
+<!DOCTYPE html>
+<html lang="en">
+
+<head>
+  <meta charset="UTF-8">
+  <meta name="viewport" content="width=device-width, initial-scale=1.0">
+  <meta http-equiv="X-UA-Compatible" content="ie=edge">
+  <link rel="stylesheet" href="/css/home.css">
+  <script src="https://cdnjs.cloudflare.com/ajax/libs/axios/0.19.0/axios.min.js"></script>
+  <title>Méditation</title>
+  {{#each css}}
+  <link rel="stylesheet" href="/css/{{this}}.css">
+  {{/each}}
+  <!-- Bootstrap -->
+  <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.4.1/css/bootstrap.min.css">
+  <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.4.1/jquery.min.js"></script>
+  <script src="https://maxcdn.bootstrapcdn.com/bootstrap/3.4.1/js/bootstrap.min.js"></script>
+  <style>
+    .carousel-inner>.item>img,
+    .carousel-inner>.item>a>img {
+      width: 50%;
+      margin: auto;
+    }
+  </style>
+</head>
+
+<body>
+
+  {{> nav}}
+
+  {{{body}}}
+
+
+
+</body>
+
+</html>
