+++ src/views/partials/nav.hbs
@@ -1,0 +1,40 @@
+<!DOCTYPE html>
+<html lang="en">
+
+<head>
+  <meta charset="UTF-8">
+  <meta name="viewport" content="width=device-width, initial-scale=1.0">
+  <meta http-equiv="X-UA-Compatible" content="ie=edge">
+  <link rel="stylesheet" href="/css/home.css">
+  <script src="https://cdnjs.cloudflare.com/ajax/libs/axios/0.19.0/axios.min.js"></script>
+  <title>Navigation bar</title>
+  <link rel="stylesheet" href="/css/home.css">
+  <!-- Bootstrap -->
+  <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.4.1/css/bootstrap.min.css">
+  <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.4.1/jquery.min.js"></script>
+  <script src="https://maxcdn.bootstrapcdn.com/bootstrap/3.4.1/js/bootstrap.min.js"></script>
+</head>
+
+<body>
+
+  <nav class="navbar navbar-inverse">
+    <div class="container-fluid">
+      <ul class="nav navbar-nav">
+        <li><a class="active" href="/">Méditation</a></li>
+        <li><a href="/contact">Formulaire de contact</a></li>
+        <li>
+          {{#unless isLoggedIn}}
+          <a href="/auth/signin" id="sign" class="button-nav">Se connecter</a>
+          {{/unless}}
+          {{#if isLoggedIn}}
+          <a href="/listeContacts">Liste des contacts</a>
+        <li><a href="/auth/logout" id="sign" class="button-nav">Se déconnecter</a></li>
+        {{/if}}
+        </li>
+      </ul>
+    </div>
+  </nav>
+
+</body>
+
+</html>
