--- views/auth/signup.hbs
@@ -1,20 +1,0 @@
- <div class="fondecran">
-<div class="form">
-<form class="form" method="post" action="/auth/signup" class="form">
-                <h1 class="text">Créer un compte</h1>
-                <label class="labelText" for="id">Identifiant</label>
-                <input class="input" name="id" type="id" required placeholder="ex: admin">
-                <label class="labelText" for="password">Mot de passe</label>
-                <input class="input" name="password" type="password" required placeholder="admin">
-                <button class="btn-light" id="send">Go !</button>
-                <hr>
-            </form>
-        
-        <div class="form">
-        <p class="text">
-            Vous avez déjà un compte ? <a class="link" href="/auth/signin">Connectez-vous !</a>
-        </p>
-        </div>
-        </div>
-        </div>
-        </div>
