--- views/listeContacts.hbs
@@ -1,11 +1,0 @@
-<div class="fondecran" id="contact">
-<h1 id="listeMessage" class="text">Liste des messages</h1>
-<section class="messages">
-    {{#each listeContacts}}
-    <p class="text">{{this.name}}</p>
-    <p class="text">{{this.firstName}}</p>
-    <p class="text">{{this.email}}</p>
-    <p class="text">{{this.message}}</p>
-    {{/each}}
-</section>
-</div>
